+++ views/backend/admin/src/components/pdk/WcShipmentLabelWrapper.vue
@@ -2,7 +2,7 @@
   <PdkBox
     v-test="AdminComponent.ShipmentLabelWrapper"
     :size="Size.Small"
-    class="mypa-mb-1 mypa-min-w-0">
+    class="mypa-mb-1 mypa-min-w-0 mypa-overflow-visible">
     <WcLoadingOverlay v-show="loading" />
 
     <slot />